Use die instead of warn & exit, simplify the script
Using die makes the program shorter & simpler, initial reason for not
using die was that it exits with 255 code which generally means: "We
died, dunno why" & I thought I knew why so I should exit with
something else.

After talking to a few people looks like it doesn't matter much, I
myself have never written a program that checks for a specific exit
code so I switched to using die.

-	print $line =~ s/\t/: /r and
-	    $total_acronyms++ if
-	    ($line =~ /^\Q${term}\E\t/i);
+	if ($line =~ /^\Q${term}\E\t/i) {
+	    print $line =~ s/\t/: /r;
+	    $total_acronyms++;
+	}

This change was made because the previous version didn't do what I
think it did. I could simply replace and with a comma & it would be
equivalent to the current version but then why not use the current
version? This looks better imo & will be easier to understand later.